requirements.txt
# local package
-e .

# external requirements
#click
#Sphinx
#coverage
#awscli
#flake8
python-dotenv
kaggle
plotly